Registering MetronomeStore in constructor instead of StoreConfig
