+++ src/js/stores/MetronomeStore.js
@@ -1,29 +1,5 @@
 import {EventEmitter} from 'events';
 
-import AppDispatcher from '../events/AppDispatcher';
-import MetronomeActions from '../events/MetronomeActions';
-import {
-  METRONOME_JOB_CREATE_ERROR,
-  METRONOME_JOB_CREATE_SUCCESS,
-  METRONOME_JOB_DELETE_ERROR,
-  METRONOME_JOB_DELETE_SUCCESS,
-  METRONOME_JOB_DETAIL_CHANGE,
-  METRONOME_JOB_DETAIL_ERROR,
-  METRONOME_JOB_UPDATE_ERROR,
-  METRONOME_JOB_UPDATE_SUCCESS,
-  METRONOME_JOB_RUN_ERROR,
-  METRONOME_JOB_RUN_SUCCESS,
-  METRONOME_JOB_STOP_RUN_SUCCESS,
-  METRONOME_JOB_STOP_RUN_ERROR,
-  METRONOME_JOB_SCHEDULE_UPDATE_ERROR,
-  METRONOME_JOB_SCHEDULE_UPDATE_SUCCESS,
-  METRONOME_JOBS_CHANGE,
-  METRONOME_JOBS_ERROR,
-  VISIBILITY_CHANGE
-} from '../constants/EventTypes';
-import Config from '../config/Config';
-import Job from '../structs/Job';
-import JobTree from '../structs/JobTree';
 import {
   REQUEST_METRONOME_JOB_CREATE_ERROR,
   REQUEST_METRONOME_JOB_CREATE_SUCCESS,
@@ -32,20 +8,44 @@ import {
   REQUEST_METRONOME_JOB_DETAIL_ERROR,
   REQUEST_METRONOME_JOB_DETAIL_ONGOING,
   REQUEST_METRONOME_JOB_DETAIL_SUCCESS,
-  REQUEST_METRONOME_JOB_UPDATE_ERROR,
-  REQUEST_METRONOME_JOB_UPDATE_SUCCESS,
   REQUEST_METRONOME_JOB_RUN_ERROR,
   REQUEST_METRONOME_JOB_RUN_SUCCESS,
-  REQUEST_METRONOME_JOB_STOP_RUN_ERROR,
-  REQUEST_METRONOME_JOB_STOP_RUN_SUCCESS,
   REQUEST_METRONOME_JOB_SCHEDULE_UPDATE_ERROR,
   REQUEST_METRONOME_JOB_SCHEDULE_UPDATE_SUCCESS,
+  REQUEST_METRONOME_JOB_STOP_RUN_ERROR,
+  REQUEST_METRONOME_JOB_STOP_RUN_SUCCESS,
+  REQUEST_METRONOME_JOB_UPDATE_ERROR,
+  REQUEST_METRONOME_JOB_UPDATE_SUCCESS,
   REQUEST_METRONOME_JOBS_ERROR,
   REQUEST_METRONOME_JOBS_ONGOING,
   REQUEST_METRONOME_JOBS_SUCCESS,
   SERVER_ACTION
 } from '../constants/ActionTypes';
-
+import AppDispatcher from '../events/AppDispatcher';
+import {
+  METRONOME_JOB_CREATE_ERROR,
+  METRONOME_JOB_CREATE_SUCCESS,
+  METRONOME_JOB_DELETE_ERROR,
+  METRONOME_JOB_DELETE_SUCCESS,
+  METRONOME_JOB_DETAIL_CHANGE,
+  METRONOME_JOB_DETAIL_ERROR,
+  METRONOME_JOB_RUN_ERROR,
+  METRONOME_JOB_RUN_SUCCESS,
+  METRONOME_JOB_SCHEDULE_UPDATE_ERROR,
+  METRONOME_JOB_SCHEDULE_UPDATE_SUCCESS,
+  METRONOME_JOB_STOP_RUN_ERROR,
+  METRONOME_JOB_STOP_RUN_SUCCESS,
+  METRONOME_JOB_UPDATE_ERROR,
+  METRONOME_JOB_UPDATE_SUCCESS,
+  METRONOME_JOBS_CHANGE,
+  METRONOME_JOBS_ERROR,
+  VISIBILITY_CHANGE
+} from '../constants/EventTypes';
+import Config from '../config/Config';
+import MetronomeActions from '../events/MetronomeActions';
+import Job from '../structs/Job';
+import JobTree from '../structs/JobTree';
+import PluginSDK from 'PluginSDK';
 import VisibilityStore from './VisibilityStore';
 
 let requestInterval;
@@ -83,6 +83,33 @@ class MetronomeStore extends EventEmitter {
       jobDetail: {},
       jobTree: {id: '/', items: []}
     };
+
+    PluginSDK.addStoreConfig({
+      store: this,
+      storeID: this.storeID,
+      events: {
+        jobCreateSuccess: METRONOME_JOB_CREATE_SUCCESS,
+        jobCreateError: METRONOME_JOB_CREATE_ERROR,
+        jobDeleteSuccess: METRONOME_JOB_DELETE_SUCCESS,
+        jobDeleteError: METRONOME_JOB_DELETE_ERROR,
+        jobDetailChange: METRONOME_JOB_DETAIL_CHANGE,
+        jobDetailError: METRONOME_JOB_DETAIL_ERROR,
+        jobUpdateSuccess: METRONOME_JOB_UPDATE_SUCCESS,
+        jobUpdateError: METRONOME_JOB_UPDATE_ERROR,
+        jobRunError: METRONOME_JOB_RUN_ERROR,
+        jobRunSuccess: METRONOME_JOB_RUN_SUCCESS,
+        jobStopRunError: METRONOME_JOB_STOP_RUN_ERROR,
+        jobStopRunSuccess: METRONOME_JOB_STOP_RUN_SUCCESS,
+        jobScheduleUpdateError: METRONOME_JOB_SCHEDULE_UPDATE_ERROR,
+        jobScheduleUpdateSuccess: METRONOME_JOB_SCHEDULE_UPDATE_SUCCESS,
+        change: METRONOME_JOBS_CHANGE,
+        error: METRONOME_JOBS_ERROR
+      },
+      unmountWhen: function () {
+        return true;
+      },
+      listenAlways: true
+    });
 
     // Handle app actions
     this.dispatcherIndex = AppDispatcher.register(({source, action}) => {
